requirements.txt
pandas
numpy
scikit-learn
python-Levenshtein
fuzzywuzzy
mlflow
kagglehub